deck: make the URL projection the deck one-way writer
Deletes the outbound sync and everything that existed to guard the loop it closed:
syncUrl, the lossy refusal, lastRepresentation, and the claim/provenance machinery.
With one writer there is nothing to tell apart.

The projection moves out of the url-handler closure into projectUrl, which the
popstate listener, deep links, boot and (next) operations all call. url-handler
drops from 603 lines to 172 and is now only listener wiring.

Adds the navigate path operations will use: currentNavigation reads the URL,
pairsForNodes converts node ids via the urlSegment stamped at graph-build time and
logs the producing extension when a node has no binding.

WARNING: intermediate state. Operations do not navigate yet, so deck changes no
longer update the URL. Not shippable until the operations are converted.
